Add files via upload
https://mp.weixin.qq.com/s/9FkCgVhygWDFWTBVhy-SpwDr. Edward Qian, chief investment officer of Pan Agora Fund, proposed the famous Risk Parity strategy. Later, this model was used in actual investment by Bridge Water Fund and became a great success.

This strategy advocates allocating risk rather than allocating assets. The traditional model hopes to control the volatility of the entire investment portfolio, while risk parity controls the relative risk, so that the risks of each asset are relatively balanced. Since the risks of the portfolio are balanced, it can theoretically withstand various risks, which is the so-called All Weather strategy.

"All Weather Trading Strategy"

The concept of the all-weather trading strategy is: "No matter what the market environment is, the investment portfolio can obtain higher risk-adjusted returns." It can be seen from the name that no matter whether the weather is stormy or clear, I remain unmoved and watch the storm with a smile.

The risk parity portfolio principle is to determine asset allocation by allocating funds so that each asset contributes equally to the risk of the entire portfolio. Fundamentally, this concept reflects the classic 1/n portfolio, where funds are equally distributed across all assets. Therefore, risk-equalized portfolios are sometimes called ERC portfolios.